+++ app/loading.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Loading = () => {
+  return (
+    <div style={{textAlign:'center', color: 'black'}}>Loading.... :(</div>
+  )
+}
+
+export default Loading
